market/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-htmx>=1.15.0
python-decouple>=3.8
